+++ CHANGELOG.md
@@ -63,6 +63,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added new resource to manage A records
 - DnsRecordAScoped
   - Added new resource to manage scoped A records
+- DnsRecordAaaa
+  - Added new resource to manage AAAA records
+- DnsRecordAaaaScoped
+  - Added new resource to manage scoped AAAA records
 - DnsRecordCname
   - Added new resource to manage CNAME records
 - DnsRecordCnameScoped